Andy's Basic Level
Walls and Floor are shades of pink, need to be replaced with art assets.
Enemies themselves have different shades of red as to indicate what they are.
Walls/Barricades are green; If we have a breakable wall code then those walls would be breakable.
Power up is placed here but can be removed if we have a better room to place them in.