Boss Animations Second Try
I think I managed to figure out how to add the boss' animations and bones to the project via importing them as a custom package.
The package itself comes with its own scene that contains the boss sprite, animations, and bones. I was able to copy the sprite from the hierarchy and paste it into the hierarchy of another scene which carried everything else over to the new scene.
Everything I imported can be found in the BossAnimations2 folder. Please let me know if this works as soon as possible so that I can apply the same technique to the turret enemy.